backend_HealthIA/requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
openai==1.64.0
python-dotenv==1.0.1
pytz==2025.1
python-multipart==0.0.9
Pillow==10.2.0
aiofiles==23.2.1
boto3==1.34.69 